.vercel/cache/wsgi_py/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==5.1.7
django-cors-headers==4.9.0
djangorestframework==3.15.2
idna==3.11
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
